Move Connection management to a separate module
